fix: aggregate workouts/events in simulate_forward_load and split llm_tools module
Bug fixes in simulate_forward_load:
- projected_day_estimate now iterates ALL workouts on a projected day and
  sums their TSS/duration. Previously only .first() was used, silently
  dropping multi-workout days.
- future_event_estimate now sums ALL events on the same date. Previously
  only the first matching event was counted.
- Main loop changed from .or_else() chain to additive logic: input +
  projected + future_event contributions are all aggregated for the same
  day. Previously sources were mutually exclusive, so a race day with
  pre-race training in input text would lose the race TSS entirely.
- Added singular/plural handling in preview_tool_arguments (1 day vs
  N days).

Module refactor:
- Split src/domain/llm_tools.rs (637 lines) into directory module:
  - mod.rs: orchestration types, run_tool_loop, tool registry, public
    tool-call mapping
  - simulate_forward_load.rs: concrete tool + helpers + tests
- No public API changes; all existing imports remain valid.

Test updates:
- Added tests for multi-workout aggregation and multi-event aggregation.
- Updated llm_adapters/training_plan.rs to expect singular 'day' in
  preview.
